move search bar back into HexView
