+++ frontend/src/HexView.svelte
@@ -44,6 +44,7 @@
 
   import { chunkList, buf2hex, hexToChar } from "./helpers.js";
   import { selectedResource, selected, settings } from "./stores.js";
+  import ResourceSearchBar from "./ResourceSearchBar.svelte";
 
   export let dataLenPromise, scrollY, resourceNodeDataMap, resources;
   let childRangesPromise = Promise.resolve(undefined);
@@ -237,6 +238,23 @@
 
     return null;
   }
+
+  let dataSearchResults = {};
+
+  // React to local data searches
+  $: {
+    const localDataSearchResults = dataSearchResults;
+
+    if (localDataSearchResults.matches !== undefined && localDataSearchResults.matches.length > 0 && localDataSearchResults.index !== undefined) {
+      dataLenPromise.then((dataLength) => {
+        $scrollY.top = localDataSearchResults.matches[localDataSearchResults.index] / dataLength
+      });
+    }
+  }
+
+  async function searchHex(query, mode) {
+    return await $selectedResource.search_data(query, mode)
+  }
 </script>
 
 {#await dataLenPromise}
@@ -257,6 +275,7 @@
         <div class="breadcrumb">
           <Breadcrumb />
         </div>
+        <ResourceSearchBar search={searchHex} bind:searchResults={dataSearchResults}/>
         <div class="hbox">
           {#await chunkDataPromise}
             <LoadingText />
